requirements.txt
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-text-splitters==0.3.11
langsmith==0.4.23
requests==2.32.5
streamlit==1.49.1